Move tests to SignTransactionWallets
