+++ src/components/SignTransactionWallets/__tests__/SignTransactionMnemonic.test.tsx
@@ -5,13 +5,12 @@ import { fTxConfig } from '@fixtures';
 import { WalletId } from '@types';
 import { translateRaw } from '@translations';
 import { WALLETS_CONFIG } from '@config';
-
-import SignTransaction from '../components/SignTransaction';
+import SignTransaction from '@features/SendAssets/components/SignTransaction';
 
 const defaultProps: React.ComponentProps<typeof SignTransaction> = {
   txConfig: {
     ...fTxConfig,
-    senderAccount: { ...fTxConfig.senderAccount, wallet: WalletId.KEYSTORE_FILE }
+    senderAccount: { ...fTxConfig.senderAccount, wallet: WalletId.MNEMONIC_PHRASE }
   },
   onComplete: jest.fn()
 };
@@ -29,7 +28,7 @@ const getHeader = (wallet: WalletId) => {
 jest.mock('ethers', () => {
   return {
     Wallet: {
-      fromEncryptedJson: jest.fn().mockImplementation(() =>
+      fromMnemonic: jest.fn().mockImplementation(() =>
         Promise.resolve({
           address: '0xfE5443FaC29fA621cFc33D41D1927fd0f5E0bB7c',
           sign: jest.fn().mockImplementation(() => Promise.resolve('txhash'))
@@ -47,20 +46,15 @@ describe('SignTransaction', () => {
     jest.resetAllMocks();
   });
 
-  test('Can handle Keystore signing', async () => {
+  test('Can render Mnemonic Phrase signing', async () => {
     const { getByText, container } = getComponent();
-    const selector = getHeader(WalletId.KEYSTORE_FILE);
+    const selector = getHeader(WalletId.MNEMONIC_PHRASE);
     expect(getByText(selector)).toBeInTheDocument();
 
-    const input = container.querySelector('input[type="file"]');
-    const file = new File(['foo'], 'keystore.json', { type: 'application/json' });
-    Object.defineProperty(input, 'files', {
-      value: [file]
-    });
-    fireEvent.change(input!);
-    const btn = container.querySelector('button[type="submit"]');
-    await waitFor(() => expect(btn!.getAttribute('disabled')).toBe(null));
-    fireEvent.click(btn!);
+    const input = container.querySelector('input');
+    fireEvent.click(input!);
+    fireEvent.change(input!, { target: { value: 'mnemonicphrase' } });
+    fireEvent.click(getByText(translateRaw('DEP_SIGNTX')));
 
     await waitFor(() => expect(defaultProps.onComplete).toBeCalledWith('txhash'));
   });
